+++ packages/comscore-analytics/CHANGELOG.md
@@ -69,6 +69,14 @@
   - @frontity/analytics@1.1.1
   - frontity@1.9.0
 
+## 0.1.1-infinite-scroll-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - frontity@1.8.1-infinite-scroll-beta.0
+  - @frontity/analytics@1.1.1-infinite-scroll-beta.0
+
 ## 0.1.0
 
 ### Minor Changes